removes unecessary function
